# swagger/swagger.yml
openapi: 3.0.0
paths:
  /campo_dinamico:
    post:
      operationId: CampoDinamicoController_post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampoDinamicoDto'
      responses:
        '201':
          description: ''
      tags: &ref_0
        - campo_dinamico
    get:
      operationId: CampoDinamicoController_getAll
      parameters:
        - name: query
          required: false
          in: query
          description: query - Filter. e.g. col1:v1,col2:v2 …
          schema:
            type: string
        - name: fields
          required: false
          in: query
          description: fields - Fields returned. e.g. col1,col2 …
          schema:
            type: string
        - name: sortby
          required: false
          in: query
          description: sortby - Sorted-by fields. e.g. col1,col2 …
          schema:
            type: string
        - name: order
          required: false
          in: query
          description: >-
            order - Order corresponding to each sortby field, if single value,
            apply to all sortby fields. e.g. desc,asc …
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: limit - Limit the size of result set. Must be an integer
          schema:
            type: string
        - name: offset
          required: false
          in: query
          description: offset - Start position of result set. Must be an integer
          schema:
            type: string
        - name: populate
          required: false
          in: query
          description: populate - show subqueries. Must be a boolean
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
  /campo_dinamico/{id}:
    get:
      operationId: CampoDinamicoController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
    put:
      operationId: CampoDinamicoController_put
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampoDinamicoDto'
      responses:
        '200':
          description: ''
      tags: *ref_0
    delete:
      operationId: CampoDinamicoController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
  /plantilla:
    post:
      operationId: PlantillaController_post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantillaDto'
      responses:
        '201':
          description: ''
      tags: &ref_1
        - plantilla
    get:
      operationId: PlantillaController_getAll
      parameters:
        - name: query
          required: false
          in: query
          description: query - Filter. e.g. col1:v1,col2:v2 …
          schema:
            type: string
        - name: fields
          required: false
          in: query
          description: fields - Fields returned. e.g. col1,col2 …
          schema:
            type: string
        - name: sortby
          required: false
          in: query
          description: sortby - Sorted-by fields. e.g. col1,col2 …
          schema:
            type: string
        - name: order
          required: false
          in: query
          description: >-
            order - Order corresponding to each sortby field, if single value,
            apply to all sortby fields. e.g. desc,asc …
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: limit - Limit the size of result set. Must be an integer
          schema:
            type: string
        - name: offset
          required: false
          in: query
          description: offset - Start position of result set. Must be an integer
          schema:
            type: string
        - name: populate
          required: false
          in: query
          description: populate - show subqueries. Must be a boolean
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
  /plantilla/{id}:
    get:
      operationId: PlantillaController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
    put:
      operationId: PlantillaController_put
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantillaDto'
      responses:
        '200':
          description: ''
      tags: *ref_1
    delete:
      operationId: PlantillaController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
  /seccion:
    post:
      operationId: SeccionController_post
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeccionDto'
      responses:
        '201':
          description: ''
      tags: &ref_2
        - seccion
    get:
      operationId: SeccionController_getAll
      parameters:
        - name: query
          required: false
          in: query
          description: query - Filter. e.g. col1:v1,col2:v2 …
          schema:
            type: string
        - name: fields
          required: false
          in: query
          description: fields - Fields returned. e.g. col1,col2 …
          schema:
            type: string
        - name: sortby
          required: false
          in: query
          description: sortby - Sorted-by fields. e.g. col1,col2 …
          schema:
            type: string
        - name: order
          required: false
          in: query
          description: >-
            order - Order corresponding to each sortby field, if single value,
            apply to all sortby fields. e.g. desc,asc …
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: limit - Limit the size of result set. Must be an integer
          schema:
            type: string
        - name: offset
          required: false
          in: query
          description: offset - Start position of result set. Must be an integer
          schema:
            type: string
        - name: populate
          required: false
          in: query
          description: populate - show subqueries. Must be a boolean
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
  /seccion/{id}:
    get:
      operationId: SeccionController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
    put:
      operationId: SeccionController_put
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeccionDto'
      responses:
        '200':
          description: ''
      tags: *ref_2
    delete:
      operationId: SeccionController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
...
info:
  title: plantillas_crud
  description: API CRUD sistema de plantillas
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
    CampoDinamicoDto:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
      required:
        - nombre
        - descripcion
    PlantillaDto:
      type: object
      properties:
        tipo_plantilla_id:
          type: string
        secciones:
          type: array
          items:
            type: string
        sistema_id:
          type: string
        grupo_id:
          type: string
        version:
          type: string
        nombre:
          type: string
        uid:
          type: boolean
        estado:
          type: boolean
        activo:
          type: boolean
        fecha_creacion:
          format: date-time
          type: string
        fecha_modificacion:
          format: date-time
          type: string
      required:
        - tipo_plantilla_id
        - secciones
        - sistema_id
        - grupo_id
        - version
        - nombre
        - uid
        - estado
        - activo
        - fecha_creacion
        - fecha_modificacion
    SeccionDto:
      type: object
      properties:
        padre_id:
          type: string
          nullable: true
        tipo_contenido_id:
          type: string
        campos_dinamicos:
          type: array
          items:
            type: string
        contenido:
          type: string
        estilos:
          type: array
          items:
            type: string
        orden:
          type: string
        activo:
          type: boolean
        fecha_creacion:
          format: date-time
          type: string
        fecha_modificacion:
          format: date-time
          type: string
      required:
        - tipo_contenido_id
        - campos_dinamicos
        - contenido
        - estilos
        - orden
        - activo
        - fecha_creacion
        - fecha_modificacion
...